[ALPS07030114] Support radio stats of tx power levels
A separate memory block passed from FW,
reported by appended to the return buffer to HAL.

The number of max levels is defined as 256 the status quo.

MTK-Commit-Id: 9b233a32263120989e2ae1c70766ffb07cd5a7a7

CR-Id: ALPS07030114
Feature: Wi-Fi Link Layer Statistics
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6060027
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_mspautosanity

+++ common/dump.c
@@ -123,9 +123,9 @@ void dumpHex(IN uint8_t *pucStartAddr, uint16_t u2Length)
 #if !DBG_DISABLE_ALL_LOG
 #define BUFSIZE 100
 	uint8_t output[BUFSIZE] = {0};
-	uint8_t i = 0;
-	uint8_t printed = 0;
-	uint8_t offset = 0;
+	uint32_t i = 0;
+	uint32_t printed = 0;
+	uint32_t offset = 0;
 
 	ASSERT(pucStartAddr);
 	LOG_FUNC("DUMPHEX ADDRESS: 0x%x, Length: %d", pucStartAddr, u2Length);
